se refactoriza la vista de RecipeDetailViewController
