+++ Cocina en Casa/Presentation/Extensiones/UIView+Ext.swift
@@ -9,7 +9,7 @@ import UIKit
 
 extension UIView {
     func addSubviews(_ views: UIView...) {
-        for view in views { addSubview(view) }
+        for view in views { self.addSubview(view) }
     }
 }
